# docs/openapi-production.yaml
openapi: 3.0.3
info:
  title: Beauty Crafter Platform API
  description: |
    Enterprise-grade beauty services platform API with comprehensive security, 
    multi-factor authentication, field-level encryption, and advanced monitoring.
    
    ## Features
    - Multi-Factor Authentication (MFA/2FA)
    - Field-level encryption for PII
    - JWT token rotation
    - API versioning
    - Rate limiting
    - Comprehensive audit logging
    - Real-time monitoring and alerting
    
    ## Security
    - All endpoints require authentication except registration
    - MFA required for sensitive operations
    - Rate limiting applied to all endpoints
    - All PII data is encrypted at rest
    - Comprehensive audit logging
    
    ## Rate Limits
    - Authentication endpoints: 5 requests per minute
    - General API endpoints: 100 requests per minute
    - File upload endpoints: 10 requests per minute
  version: 2.0.0
  contact:
    name: Beauty Crafter Support
    email: support@beautycrafter.com
    url: https://beautycrafter.com/support
  license:
    name: Proprietary
    url: https://beautycrafter.com/terms
servers:
  - url: https://api.beautycrafter.com/v2
    description: Production server
  - url: https://staging-api.beautycrafter.com/v2
    description: Staging server
  - url: http://localhost:3000/api/v2
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/mfa/setup:
    post:
      tags:
        - Authentication
      summary: Setup MFA
      description: Generate MFA secret and QR code for user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: User ID (optional, uses session if not provided)
      responses:
        '200':
          description: MFA setup successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFASetupResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/mfa/verify:
    post:
      tags:
        - Authentication
      summary: Verify MFA token
      description: Verify MFA token and optionally enable MFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MFAVerification'
      responses:
        '200':
          description: MFA verification successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFAVerificationResponse'
        '400':
          description: Invalid MFA token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /services:
    get:
      tags:
        - Services
      summary: List services
      description: Get paginated list of available services
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: category
          in: query
          schema:
            type: string
            enum: [FACIAL, MASSAGE, NAIL, HAIR, BODY, OTHER]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Services retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicesListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    MFAVerification:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          minLength: 6
          maxLength: 8
          description: MFA token from authenticator app
        enable:
          type: boolean
          default: false
          description: Whether to enable MFA after verification
...
    MFASetupResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            secret:
              type: string
              description: MFA secret for manual entry
            qrCodeUrl:
              type: string
              format: uri
              description: QR code data URL for MFA setup
            backupCodes:
              type: array
              items:
                type: string
              description: Backup codes for MFA recovery
    MFAVerificationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            isValid:
              type: boolean
            isBackupCode:
              type: boolean
...
    Service:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [FACIAL, MASSAGE, NAIL, HAIR, BODY, OTHER]
        price:
          type: number
          format: float
        duration:
          type: integer
          description: Duration in minutes
        isActive:
          type: boolean
        providerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ServicesListResponse:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: '#/components/schemas/Service'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
        code:
          type: string
        timestamp:
          type: string
          format: date-time
...